requirements.txt
asgiref==3.7.2
astroid==3.0.2
bandit==1.7.6
black==23.12.1
click==8.1.7
colorama==0.4.6
dill==0.3.7
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gitdb==4.0.11
GitPython==3.1.40
inflection==0.5.1
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pbr==6.0.0
pillow==10.2.0
platformdirs==4.1.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.7.0
smmap==5.0.1
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
